+++ pkg/monitoring/alloy/service.go
@@ -98,5 +98,6 @@ func (a *Service) ReconcileDelete(ctx context.Context, cluster *clusterv1.Cluste
 	}
 
 	logger.Info("alloy-service - ensured alloy is removed")
+
 	return nil
 }