+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add validation webhook to validate the alertmanager config before it is send to the alertmanager.
+- Add support for loading rules in a tenant in the Mimir Ruler from Workload Clusters.
 
 ## [0.24.0] - 2025-04-15