+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-- Remove loading of log-based alerts into the Loki Ruler via Alloy Rules as it is now managed by Alloy Logs.
+- Remove loading of Prometheus Rules for logs into the Loki Ruler via Alloy Rules as it is now managed by Alloy Logs.
 
 ## [0.24.0] - 2025-04-15