+++ CHANGELOG.md
@@ -10,7 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add validation webhook to validate the alertmanager config before it is send to the alertmanager.
-- Add support for loading rules in a tenant in the Mimir Ruler from Workload Clusters.
+- Add support for loading Prometheus Rules in the Mimir Ruler from workload clusters.
+
+### Changed
+
+- Load Prometheus Rules in the Mimir Ruler via Alloy Metrics instead of Alloy Rules on management clusters.
+
+### Removed
+
+- Remove loading of Prometheus Rules for logs into the Loki Ruler via Alloy Rules as it is now managed by Alloy Logs.
 
 ## [0.24.0] - 2025-04-15