+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.25.0] - 2025-04-22
+
 ### Added
 
 - Add support for loading Prometheus Rules in the Mimir Ruler from workload clusters.
@@ -429,7 +431,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initialize project and create heartbeat for the installation.
 
-[Unreleased]: https://github.com/giantswarm/observability-operator/compare/v0.24.0...HEAD
+[Unreleased]: https://github.com/giantswarm/observability-operator/compare/v0.25.0...HEAD
+[0.25.0]: https://github.com/giantswarm/observability-operator/compare/v0.24.0...v0.25.0
 [0.24.0]: https://github.com/giantswarm/observability-operator/compare/v0.23.2...v0.24.0
 [0.23.2]: https://github.com/giantswarm/observability-operator/compare/v0.23.1...v0.23.2
 [0.23.1]: https://github.com/giantswarm/observability-operator/compare/v0.23.0...v0.23.1